requirements.txt
PySide2>=5.14
PyMuPDF>=1.16
Pillow>=7.0
